прикручена бд

+++ internal/botBase/helpingMethods/createInteraction.go
@@ -3,7 +3,7 @@ package helpingMethods
 import "play_portal_bot/pkg/utils/structures"
 
 // NewInteraction creates new interaciton for user, optPrice and optData is optional
-func NewInteraction(interactionType string, chatID int64, optPrice int, optData []string) {
+func NewInteraction(interactionType string, chatID int64, optPrice float64, optData []string) {
 	delete(structures.UserStates, chatID)
 	structures.UserStates[chatID] = &structures.UserInteraction{
 		IsInteracting: true,
